+++ src/propagators/j2.jl
@@ -329,7 +329,8 @@ elements represented by a set of position vectors `vr_i` [m] and a set of veloci
     (**Default**: 1e-7)
 - `max_iterations::Int`: Maximum number of iterations allowed for the least-square fitting.
     (**Default**: 50)
-- `mean_elements_epoch::Number`: Epoch for the fitted mean elements.
+- `mean_elements_epoch::Union{Number, DateTime}`: Epoch of the fitted mean elements,
+    represented by a Julian Day [UTC] or a `DateTime` [UTC].
     (**Default**: vjd[end])
 - `verbose::Bool`: If `true`, the algorithm prints debugging information to `stdout`.
     (**Default**: true)
@@ -480,7 +481,8 @@ elements represented by a set of position vectors `vr_i` [m] and a set of veloci
     (**Default**: 1e-7)
 - `max_iterations::Int`: Maximum number of iterations allowed for the least-square fitting.
     (**Default**: 50)
-- `mean_elements_epoch::Number`: Epoch for the fitted mean elements.
+- `mean_elements_epoch::Union{Number, DateTime}`: Epoch of the fitted mean elements,
+    represented by a Julian Day [UTC] or a `DateTime` [UTC].
     (**Default**: vjd[end])
 - `verbose::Bool`: If `true`, the algorithm prints debugging information to `stdout`.
     (**Default**: true)